added site_ctl

+++ site_ctl/site_ctl.py
@@ -12,63 +12,36 @@ $ site_ctl help
 
 Prefix = "CE_cfg"
 
-Params = [
-    "ce_disabled",
-    "sever",
-    "server_root",
-    "timeout",
-    "roots",
-    "ignore_list",
-    "nworkers",
-    "max_dark_fraction",
-    "max_missing_fraction"
+Params = [Prefix + "." + n for n in 
+    [
+        "ce_disabled",
+        "sever",
+        "server_root",
+        "timeout",
+        "roots",
+        "ignore_list",
+        "nworkers",
+        "max_dark_fraction",
+        "max_missing_fraction"
+    ]
 ]
-
-Lists = [
-    "roots",
-    "ignore_list"
-]
-
-def internal_name(name):
-    if not name.startswith(Prefix + '.'):
-        name = Prefix + '.' + name
-    return name
-
-def external_name(name):
-    if name.startswith(Prefix + '.'):
-        name = name[len(Prefix) + 1:]
-    return name
-
-def internal_to_external(config):
-    config = {external_name(name): value for name, value in config.items() if name in Params}
-    for name in Lists:
-        if name in config:
-            config["name"] = config["name"].split('/')
-    return config
-
-def external_to_internal(config):
-    config = config.copy()
-    for name in Lists:
-        if name in config:
-            config[name] = ','.join(config[name])
-    return {internal_name(name): value for name, value in config.items() if name in Params}
 
 def read_config(rse):
     client = RSEClient()
     params = client.list_rse_attributes(rse)
     config = {name: value for name, value in params.items() if name in Params}
-    return internal_to_external(config)
+    assert config.get(Prefix + "." + "ce_disabled", "false") in ("true", "false")
+    return config
 
 def write_config(rse, config):
-    config = external_to_internal(config)
+    assert config.get(Prefix + "." + "ce_disabled", "false") in ("true", "false")
     client = RSEClient()
     existing_config = client.list_rse_attributes(rse)
     for name in existing_config:
-        if name not in config:
+        if name in Params and name not in config:
             client.delete_rse_attribute(rse, name)
     for name, value in config.items():
-        client.add_rse_attribute(rse, name, value)
-    return config
+        client.add_rse_attribute(rse, name, str(value))
 
 #
 #
@@ -95,7 +68,7 @@ def do_set(rse, name, value):
 
 def do_get(rse, name):
     config = read_config(rse)
-    print(name, config.get(name, ""))
+    print(name, config.get(name, "-"))
 
 def main():
     cmd, args = sys.argv[1], sys.argv[2:]
